Remove all FFTW precision dependencies
